// src/models/sports-table-data.ts
// Generated by https://quicktype.io

// https://push.api.bbci.co.uk/batch?t=/data/bbc-morph-sport-tables-data/competition/championship/sport/football/version/2.0.2?timeout=5

export interface SportsTableData {
  meta: SportsTableDataMeta;
  payload: Payload[];
}

export interface SportsTableDataMeta {
  pollFrequencyInMilliseconds: number;
}

export interface Payload {
  meta: PayloadMeta;
  body: Body;
}

export interface Body {
  sportTables: SportTables;
}

export interface SportTables {
  title: string;
  tables: Table[];
}

export interface Table {
  group: Group;
  live: boolean;
  headings: Heading[];
  rows: Row[];
  footer: Footer[];
}

export interface Footer {
  cells: FooterCell[];
}

export interface FooterCell {
  td: PurpleTd;
}

export interface PurpleTd {
  colspan: number;
  time: string;
  text: string;
  styles: string;
}

export interface Group {
  name: null;
}

export interface Heading {
  cells: HeadingCell[];
}

export interface HeadingCell {
  th: Th;
}

export interface Th {
  text: null | string;
  live?: boolean;
  styles?: string;
  abbr?: string;
  abbrStyles?: string;
}

export interface Row {
  rowStyles: RowStyles;
  cells: RowCell[];
}

export interface RowCell {
  td: TableCell;
}

export interface TableCell {
  text?: TextEnum | number;
  icon?: Icon;
  styles?: string;
  abbrLink?: AbbrLink;
  form?: Form[];
}

export interface AbbrLink {
  link: string;
  text: string;
  abbr: string;
  abbrStyles: AbbrStyles;
}

export enum AbbrStyles {
  SPUAbbrOnSPUAbbrOffM = "sp-u-abbr-on sp-u-abbr-off@m",
}

export interface Form {
  styles: Styles;
  text: string;
  result: Result;
}

export enum Result {
  D = "D",
  L = "L",
  W = "W",
}

export enum Styles {
  SPCTeamFormSPCTeamFormDrewGelMinion = "sp-c-team-form sp-c-team-form--drew gel-minion",
  SPCTeamFormSPCTeamFormLostGelMinion = "sp-c-team-form sp-c-team-form--lost gel-minion",
  SPCTeamFormSPCTeamFormWonGelMinion = "sp-c-team-form sp-c-team-form--won gel-minion",
}

export enum Icon {
  SPCMovement = "sp-c-movement",
  SPCMovementSPCMovementDown = "sp-c-movement sp-c-movement--down",
  SPCMovementSPCMovementUp = "sp-c-movement sp-c-movement--up",
}

export enum TextEnum {
  TeamHasMovedDown = "team has moved down",
  TeamHasMovedUp = "team has moved up",
  TeamHasnTMoved = "team hasn't moved",
}

export enum RowStyles {
  Empty = " ",
  GsOTableRowBreak = "gs-o-table__row--break ",
}

export interface PayloadMeta {
  responseCode: number;
  hash: string;
  template: string;
}